+++ release-installer/tn-code-release-installer/tn-code-release-installer.php
@@ -3,7 +3,7 @@
  * Plugin Name: TN Code Release Installer
  * Plugin URI: https://github.com/cchatterton/tn-wp-migrate-code-diff/releases/latest
  * Description: Installs manual code release packages created by TN WP Migrate Code Diff.
- * Version: 0.4.0
+ * Version: 0.4.1
  * Requires at least: 5.2
  * Requires PHP: 5.6
  * Update URI: https://github.com/cchatterton/tn-wp-migrate-code-diff
@@ -16,7 +16,7 @@ if (!defined('ABSPATH')) {
     exit;
 }
 
-define('TNCRI_VERSION', '0.4.0');
+define('TNCRI_VERSION', '0.4.1');
 define('TNCRI_PLUGIN_FILE', __FILE__);
 define('TNCRI_PLUGIN_DIR', plugin_dir_path(__FILE__));
 define('TNCRI_PLUGIN_URL', plugin_dir_url(__FILE__));